updated python

+++ .github/scripts/process_and_publish.py
@@ -19,6 +19,23 @@ def clone_repo(repo_url, target_dir):
     else:
         print(f"Directory {target_dir} exists, assuming repository already cloned.")
 
+# def copy_chart_to_charts_dir(source_dir, chart_path):
+#     """
+#     Copies the chart from the specified path within the cloned repo to the
+#     charts directory, creating a subdirectory named after the last segment of the path.
+#     """
+#     chart_name = chart_path.split('/')[-1]
+#     dest_path = os.path.join(charts_dir, chart_name)
+
+#     # Calculate the full source path
+#     full_source_path = os.path.join(source_dir, chart_path)
+
+#     if os.path.exists(dest_path):
+#         shutil.rmtree(dest_path)
+#     shutil.copytree(full_source_path, dest_path)
+
+#     print(f"Copied {full_source_path} to {dest_path}")
+
 def copy_chart_to_charts_dir(source_dir, chart_path):
     """
     Copies the chart from the specified path within the cloned repo to the
@@ -29,6 +46,10 @@ def copy_chart_to_charts_dir(source_dir, chart_path):
 
     # Calculate the full source path
     full_source_path = os.path.join(source_dir, chart_path)
+
+    if not os.path.exists(full_source_path):
+        print(f"Source path does not exist: {full_source_path}")
+        return
 
     if os.path.exists(dest_path):
         shutil.rmtree(dest_path)
